feat(avatar): implement GET /account/{uuid}/avatar
Either the avatar associated with the given `{uuid}` is returned, or the
default avatar is return in the case that there does not exist an
associated avatar.

Avatars are stored as PNG images under the `data/avatars` directory. The
default avatar `data/avatar/default.png` we used is granted **permission
for free commerical use** but copyright attribution is hereby given:

COPYRIGHT ATTRIBUTION

> Author: Anna Litviniuk
> Source: https://www.iconfinder.com/icons/403017/anonym_avatar_default_head_person_unknown_user_icon