perf(avatar): delegate avatar read to thread pool
Upon inspect, I found that `actix-files::NamedFile::open` is *not*
asynchronous and directly calls `std::fs::File::open` to read the avatar
file, which is a blocking function.

On one request, delegating to thread pool could be a little bit slower
due to some overhead, but when multiple incoming requests occur the
blocking will become significant if we don't delegate the blocking file
read to the thread pool.